requirements-security.txt
# Security Dependencies for Phase 0
# Install with: pip install -r requirements-security.txt

# Authentication & Session Management
Flask-Login==0.6.3
bcrypt==4.1.1

# Input Validation & Sanitization
marshmallow==3.20.1
email-validator==2.1.0
bleach==6.1.0

# CSRF Protection
Flask-WTF==1.2.1
WTForms==3.1.1

# Security Headers
Flask-Talisman==1.1.0

# Structured Logging
python-json-logger==2.0.7

# Environment Variables
python-dotenv==1.0.0

# Database (if not already installed)
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
psycopg2-binary==2.9.9

# Rate Limiting
Flask-Limiter==3.5.0
redis==5.0.1

# Password Strength Validation
zxcvbn==4.4.28
